feat: 비회원 좋아요 관련 구현 및 수정(#WA-172)

+++ src/main/java/io/wisoft/wasabi/domain/board/dto/ReadBoardResponse.java
@@ -10,7 +10,7 @@ public record ReadBoardResponse(
         String content,
         Writer writer,
         LocalDateTime createdAt,
-        int likeCount,
+        long likeCount,
         int views,
         boolean isLike,
         String tag) {
